+++ src/main/scala/edu/ie3/simona/agent/em/EmDataCore.scala
@@ -8,7 +8,7 @@ package edu.ie3.simona.agent.em
 
 import edu.ie3.simona.agent.em.EmAgent.Actor
 import edu.ie3.simona.agent.em.FlexCorrespondenceStore.WithTime
-import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ApparentPower
+import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ComplexPower
 import edu.ie3.simona.exceptions.CriticalFailureException
 import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage._
 import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
@@ -163,7 +163,7 @@ object EmDataCore {
 
     /** Returns relevant results for all connected agents.
       */
-    def getResults: Iterable[ApparentPower] =
+    def getResults: Iterable[ComplexPower] =
       correspondences.store.values.flatMap(_.receivedResult.map(_.get))
 
   }